+++ src/Controller/MonitoringController.php
@@ -24,8 +24,9 @@ class MonitoringController extends AbstractController
 {
     private const ALLOWED_KEYS = [
         'events',
-        'hosts',
         'events/latest',
+        'events/types',
+        'hosts',
         'users',
         'wallet',
     ];
@@ -43,8 +44,8 @@ class MonitoringController extends AbstractController
 
         try {
             $license = $licenseReader->read($licenseKey);
-        } catch (OutdatedLicense) {
-            throw new UnprocessableEntityHttpException('License is outdated');
+        } catch (OutdatedLicense $exception) {
+            throw new UnprocessableEntityHttpException($exception->getMessage());
         } catch (ServiceNotPresent $exception) {
             throw new HttpException(Response::HTTP_GATEWAY_TIMEOUT, $exception->getMessage());
         } catch (UnexpectedResponseException $exception) {